+++ Market/Source/ThreadSafeRNG.h
@@ -15,7 +15,7 @@ std::mutex my_rng_mutex; // a mutex to guard my_rng
 
 
 // This is the function to call if you want a random number in the interval [0,1)
-double RANDOM() {
+float RANDOM() {
     std::lock_guard<std::mutex> lock(my_rng_mutex);
     return my_unif_real_dist(my_rng);
     // mutex is released when lock goes out of scope
